Initial Commenting and .obj mixup
I have begun the long process of commenting the files which were given to us. I aim to keep it simple and easy to read. If there are any issues or things you want me to change about the wa comments are handles just let me know.

Following this a list of coding standards shoudl be creted to keep continuity throughout this small project.

I have also adjusted the GitIgnore to no longer ignore the .obj files so now the engine should work properly if loaded from this repo.

At a later time I will finish the commenting process and we can start working. Although, a game idea must be made first giving plenty of time for the commenting to be finished.