Gravity Component
I have added in a component which if the object has a transform component then the object will slowly move down. Once fully implemented this value will need to be tweeked.

I will also need to add in some collision detection with the flat plane of the ground so that the object will no longer move if touching a ground object.